requirements.txt
asgiref==3.5.0
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-link-header-pagination==0.1.1
importlib-metadata==4.11.2
Markdown==3.3.6
pytz==2021.3
sqlparse==0.4.2
zipp==3.7.0
